Backend/requirements.txt
Flask
Flask-Cors
param
langchain
openai
pypdf
chromadb
tiktoken
langchain-openai
python-dotenv
langchain-community